+++ src/action-creators/userAuthenticated.ts
@@ -1,7 +1,7 @@
 import { getUserRef } from '../util'
 import { authenticate, error, status } from '../action-creators'
 import { Thunk, User } from '../types'
-import { storage } from '../util/localStorage'
+import { storage } from '../util/storage'
 
 /** Updates local state with newly authenticated user. */
 const userAuthenticated =
