+++ src/test-helpers/createTestApp.tsx
@@ -15,7 +15,7 @@ import AppComponent from '../components/AppComponent'
 import ErrorBoundaryContainer from '../components/ErrorBoundaryContainer'
 import TouchMonitor from '../components/TouchMonitor'
 import testTimer from './testTimer'
-import { storage } from '../util/localStorage'
+import { storage } from '../util/storage'
 
 /**
  * Test App.
